Move all CLI-related code to `cli.rs`
